+++ hive_generator/example/lib/types.g.dart
@@ -205,6 +205,46 @@ class ConstructorDefaultsAdapter extends TypeAdapter<ConstructorDefaults> {
           typeId == other.typeId;
 }
 
+class NullableTypesAdapter extends TypeAdapter<NullableTypes> {
+  @override
+  final int typeId = 7;
+
+  @override
+  NullableTypes read(BinaryReader reader) {
+    final numOfFields = reader.readByte();
+    final fields = <int, dynamic>{
+      for (int i = 0; i < numOfFields; i++) reader.readByte(): reader.read(),
+    };
+    return NullableTypes(
+      a: (fields[0] as num?)?.toInt(),
+      b: fields[1] as String?,
+      c: fields[2] as bool?,
+    );
+  }
+
+  @override
+  void write(BinaryWriter writer, NullableTypes obj) {
+    writer
+      ..writeByte(3)
+      ..writeByte(0)
+      ..write(obj.a)
+      ..writeByte(1)
+      ..write(obj.b)
+      ..writeByte(2)
+      ..write(obj.c);
+  }
+
+  @override
+  int get hashCode => typeId.hashCode;
+
+  @override
+  bool operator ==(Object other) =>
+      identical(this, other) ||
+      other is NullableTypesAdapter &&
+          runtimeType == other.runtimeType &&
+          typeId == other.typeId;
+}
+
 class Enum1Adapter extends TypeAdapter<Enum1> {
   @override
   final int typeId = 3;