+++ hive_generator/lib/src/class_builder.dart
@@ -1,6 +1,5 @@
 import 'dart:typed_data';
 
-import 'package:analyzer/dart/constant/value.dart';
 import 'package:analyzer/dart/element/nullability_suffix.dart';
 import 'package:analyzer/dart/element/type.dart';
 import 'package:hive_ce/hive.dart';
@@ -68,13 +67,7 @@ class ClassBuilder extends Builder {
         if (param.isNamed) {
           code.write('${param.name}: ');
         }
-        code.write(
-          _value(
-            param.type,
-            'fields[${field.index}]',
-            field.defaultValue,
-          ),
-        );
+        code.write(_value(param.type, field));
         code.writeln(',');
         fields.remove(field);
       }
@@ -86,13 +79,7 @@ class ClassBuilder extends Builder {
     // as initializing formals. We do so using cascades.
     for (final field in fields) {
       code.write('..${field.name} = ');
-      code.writeln(
-        _value(
-          field.type,
-          'fields[${field.index}]',
-          field.defaultValue,
-        ),
-      );
+      code.writeln(_value(field.type, field));
     }
 
     code.writeln(';');
@@ -100,10 +87,25 @@ class ClassBuilder extends Builder {
     return code.toString();
   }
 
-  String _value(DartType type, String variable, DartObject? defaultValue) {
+  String _value(DartType type, AdapterField field) {
+    final variable = 'fields[${field.index}]';
     final value = _cast(type, variable);
-    if (defaultValue?.isNull != false) return value;
-    return '$variable == null ? ${constantToString(defaultValue!)} : $value';
+
+    final annotationDefaultIsNull = field.annotationDefault?.isNull ?? true;
+    final constructorDefaultIsNull = field.constructorDefault == null;
+
+    final String? defaultValue;
+    if (!annotationDefaultIsNull) {
+      defaultValue = constantToString(field.annotationDefault);
+    } else if (!constructorDefaultIsNull) {
+      defaultValue = field.constructorDefault;
+    } else {
+      defaultValue = null;
+    }
+
+    if (defaultValue == null) return value;
+
+    return '$variable == null ? $defaultValue : $value';
   }
 
   String _cast(DartType type, String variable) {
@@ -121,7 +123,7 @@ class ClassBuilder extends Builder {
     } else if (type.isDartCoreDouble) {
       return '($variable as num$suffix)$suffix.toDouble()';
     } else {
-      return '$variable as ${_displayString(type)}';
+      return '$variable as ${type.getDisplayString()}';
     }
   }
 
@@ -152,7 +154,7 @@ class ClassBuilder extends Builder {
 
       return '$suffix.map((e) => ${_cast(arg, 'e')})$cast';
     } else {
-      return '$suffix.cast<${_displayString(arg)}>()';
+      return '$suffix.cast<${arg.getDisplayString()}>()';
     }
   }
 
@@ -165,8 +167,8 @@ class ClassBuilder extends Builder {
       return '$suffix.map((dynamic k, dynamic v)=>'
           'MapEntry(${_cast(arg1, 'k')},${_cast(arg2, 'v')}))';
     } else {
-      return '$suffix.cast<${_displayString(arg1)}, '
-          '${_displayString(arg2)}>()';
+      return '$suffix.cast<${arg1.getDisplayString()}, '
+          '${arg2.getDisplayString()}>()';
     }
   }
 
@@ -212,16 +214,8 @@ String _accessorSuffixFromType(DartType type) {
 /// Suffix to use when casting a value to [type].
 /// $variable as $type$suffix
 String _suffixFromType(DartType type) {
-  if (type.nullabilitySuffix == NullabilitySuffix.star) {
-    return '';
-  }
-  if (type.nullabilitySuffix == NullabilitySuffix.question) {
-    return '?';
-  }
-  return '';
-}
-
-String _displayString(DartType e) {
-  final suffix = _suffixFromType(e);
-  return '${e.getDisplayString()}$suffix';
+  return switch (type.nullabilitySuffix) {
+    NullabilitySuffix.question => '?',
+    _ => '',
+  };
 }