+++ hive_generator/CHANGELOG.md
@@ -1,3 +1,10 @@
+## 1.5.0
+
+- Supports constructor parameter default values
+- No longer generates the `HiveRegistrar` if there are no adapters
+- Removes unnecessary print statement in `HiveRegistrar` generator
+- Bumps `analyzer` to `^6.5.0` to deal with deprecations
+
 ## 1.4.0
 
 - Adds a generator to create a `HiveRegistrar` extension that allows registration of all generated `TypeAdapters` in one call